% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{imputation}
\alias{imputation}
\title{Pattern selection by SVD}
\usage{
imputation(target, location_corr)
}
\arguments{
\item{location}{the formatted location information}
}
\value{
a location preview plot
}
\description{
Make a location preview plot for user to directly see the distribution of sample slots
}
