% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_convert.R
\name{loc_convert}
\alias{loc_convert}
\title{Location Formatting Function}
\usage{
loc_convert(location)
}
\arguments{
\item{location}{the location information of each micro-slot}
}
\value{
a new format that the codes are easy to deal with
}
\description{
This function converts the location information to an coding-tractable format
}
