% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyDistriPlot.R
\name{MyDistriPlot}
\alias{MyDistriPlot}
\title{Distribution map generation of proteins}
\usage{
MyDistriPlot(
  df,
  location,
  target_name,
  limit,
  scale = TRUE,
  upperL = 2,
  lowerL = -2,
  color1,
  color2,
  color3,
  transparent = FALSE,
  mask = FALSE,
  Legend = TRUE,
  g = g
)
}
\arguments{
\item{df}{the quantitative protein result}

\item{location}{the original location information}

\item{target_name}{the protein in the map}

\item{limit}{a logical value, if setting the value limits of protein abundance/ratio in the distribution map}

\item{scale}{a logical value, if converting data to z-score}

\item{upperL}{the upperbound of the protein abundance/ratio the map can show}

\item{lowerL}{the lowerbound of the protein abundance/ratio the map can show}

\item{color1}{the lower color, color name and HTML code are both acceptable}

\item{color2}{the middle color, color name and HTML code are both acceptable}

\item{color3}{the upper color, color name and HTML code are both acceptable}

\item{transparent}{a logical value, if the map is with a transparent background while downloading}

\item{mask}{a logical value, if adding a brain-shape cover on the map}

\item{Legend}{a logical value, if showing the legend}

\item{g}{the brain-shape cover}
}
\value{
a protein distribution map
}
\description{
Distribution map generation of proteins
}
