% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyPreview.R
\name{MyPreview}
\alias{MyPreview}
\title{Location Preview Plot}
\usage{
MyPreview(location)
}
\arguments{
\item{location}{the formatted location information}
}
\value{
a location preview plot
}
\description{
Make a location preview plot for user to directly see the map of sample slots
}
