% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SClustering.R
\name{SClustering}
\alias{SClustering}
\title{Spectral Clustering on filtered data}
\usage{
SClustering(df_filtered)
}
\arguments{
\item{df_filtered}{the filtered protein dataset}
}
\value{
a PCA plot colored by label from spectral clustering, and a dataset containing cluster label.
}
\description{
Spectral Clustering on filtered data
}
