% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationPlots.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation plots}
\usage{
correlation_plot(
  df,
  target_name,
  type = "pearson",
  location,
  limit,
  upperL = 2,
  lowerL = -2,
  color1,
  color2,
  color3,
  transparent = FALSE,
  mask = FALSE,
  g = g
)
}
\arguments{
\item{df_filtered}{the filtered protein dataset}
}
\value{
a PCA plot colored by label from spectral clustering, and a dataset containing cluster label.
}
\description{
Correlation plots
}
